requirements.txt
alembic==1.10.2
Babel==2.8.0
blinker==1.4
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
filelock==3.9.0
Flask==2.2.3
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
httplib2==0.20.2
importlib-metadata==4.6.4
itsdangerous==2.1.0
jeepney==0.7.1
Jinja2==3.0.3
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Mako==1.2.4
MarkupSafe==2.1.2
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
platformdirs==3.1.1
psycopg2-binary==2.9.5
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
python-apt==2.3.0+ubuntu2.1
pytz==2022.1
PyYAML==5.4.1
SecretStorage==3.3.1
simplejson==3.17.6
six==1.16.0
SQLAlchemy==2.0.5.post1
systemd-python==234
typing_extensions==4.5.0
ubuntu-advantage-tools==27.12
ufw==0.36.1
unattended-upgrades==0.1
virtualenv==20.20.0
wadllib==1.3.6
Werkzeug==2.2.3
WTForms==3.0.1
zipp==1.0.0
